refactor: break out everything into modules by tab
